publisher-v2/requirements.txt
GitPython>=3.1.0
inflection>=0.5.1
Pillow>=9.0.0
pytz>=2021.3
PyYAML>=6.0
titlecase>=2.3
